updated dummy data

+++ app/src/main/java/com/example/android/quakereport/Earthquake.java
@@ -1,18 +1,18 @@
 package com.example.android.quakereport;
 
 public class Earthquake {
-    private float magnitude;
+    private double magnitude;
     private String location;
     private String date;
 
-    public Earthquake(float magnitude, String location, String date){
+    public Earthquake(double magnitude, String location, String date){
         this.magnitude = magnitude;
         this.location = location;
         this.date = date;
     }
 
 
-    public float getMagnitude() {
+    public double getMagnitude() {
         return magnitude;
     }
